Move `PollResultsDialogFragment` to a custom package
